fmfaucet/requirements.txt
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
gunicorn>=20.1
mock==5.1.0
python-dotenv==1.0.1
redis==5.0.8
web3==7.2.0
